Add files via upload
JSX and CSS files for the Menu and Map components. The Menu component uses the the MenuItem component. The Menu component and the Map component are bundled in the Interactive map component to have the Menu on the left side of the map. The leaflet2.css file updates the default css of the leaflet library to customize the pin components and the map clusters.