Code for the Node.js/ Express server
This server listens for a get request from the front end to retrieve the processed and predicted data from the database. The front end issues a get request "/getData" and the server responds with a json file containing the data fields.  Right now the server is configured to run on localhost port 3001. This will likely change if we move the server to some sort of cloud hosting service. The client can request the data using the url: "'http://localhost:3001/getData".