updated with api call to database
The App component make a get request via the fetch method to 'http://localhost:3001/getData'. The data that is received is a json file containing an array of objects for each disaster. The array is then passed into the InteractiveMap component as a prop for the child components to use.